requirements.txt
streamlit
google-generativeai
pillow
speechrecognition
gtts
translate
